+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - function `displayUsing`
+- function `showUsing`
 
 ## [1.2.0] (2021-04-03)